requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
PyPDF2
python-pptx
docx2txt
python-dotenv
